Move nested class in LookupJoinOperatorFactory as standlone class
Nested class PerLifespanDataManager was responsible for managing the
lifecycle (reference count, destroy) of data that's shared between join
build and probe. The class is not only useful for Hash Lookup join,
but also makes sense for Nested Loop join. The class is moved into
a standalone class, and renamed to JoinBridgeLifecycleManager.
